Login e cadastro de funcionarios

+++ public/index.php
@@ -1,16 +1,28 @@
 <?php
 // Inicia a sessão (você vai precisar disso para o login do vendedor depois)
+ini_set('display_errors', 1);
+error_reporting(E_ALL);
 session_start();
 
 require_once __DIR__ . '/../app/Config/Root.php';
 require_once ROOT_PATH . '/app/Database/Conexao.php';
 require_once ROOT_PATH . '/app/Model/Veiculo.php';
-require_once ROOT_PATH . '/app/Repository/Veiculos/VeiculoRepository.php';
+require_once ROOT_PATH . '/app/Repository/Veiculo/VeiculoRepository.php';
 require_once ROOT_PATH . '/app/Controller/Veiculo/VeiculoController.php';
+require_once ROOT_PATH . '/app/Model/User.php';
+require_once ROOT_PATH . '/app/Repository/Users/UserRepository.php';
+require_once ROOT_PATH . '/app/Service/UserService.php';
+require_once ROOT_PATH . '/app/Controller/User/UserController.php';
 require_once ROOT_PATH . '/app/Helpers/Validadores.php';
 
 // Iniciar o Controller
 $controller = new VeiculoController($pdo);
+
+
+// User
+$userRepository = new UserRepository($pdo);
+$userService    = new UserService($userRepository);
+$userController = new UserController($userService);
 
 // Pega qual página o usuário quer acessar da URL. Ex: index.php?pagina=admin
 // Se ele não digitar nada, a página padrão será a 'home'
@@ -18,6 +30,12 @@ $pagina = isset($_GET['pagina']) ? $_GET['pagina'] : 'home';
 
 // Caminho base para a pasta onde estão suas telas
 $caminho_views = '../app/Views/';
+
+
+function estaLogado(): bool
+{
+    return isset($_SESSION['user_id']);
+}
 
 // O Roteador: decide qual arquivo carregar
 switch ($pagina) {
@@ -53,19 +71,64 @@ switch ($pagina) {
     // ÁREA DO VENDEDOR (ADMIN)
     // ==========================================
     case 'login':
+        if (estaLogado()) {
+            header('Location: ?pagina=painel');
+            exit;
+        }
+
         require_once $caminho_views . 'admin/login.php';
         break;
+    // CADASTRO USER
+    case 'register-admin':
+        require_once $caminho_views . 'admin/register_admin.php';
+        break;
+
+    case 'processar-login':
+        if (
+            $_SERVER['REQUEST_METHOD'] === 'POST'
+            && $userController->login()
+        ) {
+            header('Location: ?pagina=painel');
+            exit;
+        }
+        header('Location: ?pagina=login&erro_login=1');
+        exit;
+
+    case 'processar-cadastro-admin':
+        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+            $cadastrado = $userController->cadastrarUser();
+        }
+        if ($cadastrado) {
+            header('Location: ?pagina=login');
+            exit;
+        }
+        header('Location: ?pagina=register-admin&erro_register=1');
+        exit;
+
 
     case 'painel':
-        // No futuro, aqui você verifica se o vendedor está logado usando $_SESSION
+        if (!estaLogado()) {
+            header('Location: ?pagina=login');
+            exit;
+        }
+
         require_once $caminho_views . 'admin/painel.php';
         break;
 
+    // CADASTRAR VEICULO
     case 'cadastrar':
+        if (!estaLogado()) {
+            header('Location: ?pagina=login');
+            exit;
+        }
         require_once $caminho_views . 'admin/cadastrar.php';
         break;
 
     case 'editar':
+        if (!estaLogado()) {
+            header('Location: ?pagina=login');
+            exit;
+        }
         require_once $caminho_views . 'admin/editar.php';
         break;
 
@@ -125,9 +188,9 @@ switch ($pagina) {
             exit;
         }
         break;
-    
+
     case 'processar_compra':
-        if($_SERVER['REQUEST_METHOD'] == 'POST') {
+        if ($_SERVER['REQUEST_METHOD'] == 'POST') {
             // Se for Compra, valida o WhatsApp antes de continuar
             if (!validar_whatsapp($_POST['telefone'])) {
                 // WhatsApp inválido: volta pra tela do carro com erro
@@ -151,6 +214,10 @@ switch ($pagina) {
         break;
 
     case 'processar_cadastro':
+        if (!estaLogado()) {
+            header('Location: ?pagina=login');
+            exit;
+        }
         // Só aceita se vier de um formulário via POST
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $controller->cadastrar();
@@ -161,6 +228,10 @@ switch ($pagina) {
         break;
 
     case 'processar_edicao':
+        if (!estaLogado()) {
+            header('Location: ?pagina=login');
+            exit;
+        }
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             // Pega o ID que mandamos escondido no formulário
             $id = $_POST['id'];
@@ -171,6 +242,10 @@ switch ($pagina) {
         break;
 
     case 'deletar':
+        if (!estaLogado()) {
+            header('Location: ?pagina=login');
+            exit;
+        }
         $id = isset($_GET['id']) ? $_GET['id'] : null;
         if ($id) {
             $controller->delete($id);
